+++ tooling/cli/src/helpers/mod.rs
@@ -8,6 +8,7 @@ pub mod cargo_manifest;
 pub mod config;
 pub mod flock;
 pub mod framework;
+pub mod fs;
 pub mod npm;
 #[cfg(target_os = "macos")]
 pub mod pbxproj;