+++ core/tauri-runtime/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## \[2.0.0-rc.5]
+
+### Dependencies
+
+- Upgraded to `tauri-utils@2.0.0-rc.5`
+
 ## \[2.0.0-rc.4]
 
 ### Dependencies